expense_elt/requirements.txt
pdfplumber>=0.11
duckdb>=1.0
typer>=0.12
rapidfuzz>=3.0
pyyaml>=6.0
rich>=13.0
python-dateutil>=2.9
streamlit>=1.35
plotly>=5.0
pandas>=2.0
monopoly-core>=0.9
fastapi>=0.115
uvicorn[standard]>=0.32
anthropic>=0.40
openai>=1.50
google-auth>=2.0
PyJWT>=2.0
slowapi>=0.1.9
python-dotenv>=1.0
